requirements.txt
qieyun-python==0.13.1
